lint and recommender path for prod corrected
