+++ app/src/app/angular2-material-meteor-accounts-ui/InjectUserAnnotation.ts
@@ -1,0 +1,4 @@
+export class InjectUserAnnotation {
+    constructor(public propName: string = 'user') {
+    }
+}
