+++ app/src/app/angular2-material-meteor-accounts-ui/accounts-module.ts
@@ -8,7 +8,7 @@ import { MatInputModule } from '@angular/material/input';
 import { MatMenuModule } from '@angular/material/menu';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 
-import { AuthGuard } from './annotations';
+import { AuthGuard } from "./AuthGuard";
 import { LoginButtons } from './loginButton/login-buttons';
 import { LoginDialog } from './loginDialog/login-dialog';
 import { SignupDialog } from './signupDialog/signup-dialog';
