+++ app/src/environments/environment.prod.ts
@@ -2,7 +2,6 @@ export const environment = {
     production: true,
     recommender:{
         base: "/recommender/batchRecommender",
-        port: '',
     } ,
     guestUserName: window["env"]["guser"] || "guest",
     guestPassword: window["env"]["gpass"] || "12345",
